fix(php8.3): eliminate all E_DEPRECATED in captcha generation
Implicit float->int precision loss + dynamic property creation deprecations
fired thousands of log lines per captcha on PHP 8.2/8.3. Fixes:
- #[\AllowDynamicProperties] on class Securimage (dynamic props $code_entered/$correct_code/etc)
- (int) casts: drawLines imagefilledrectangle, distortion imagecolorat (per-pixel),
  pole placement (% modulo + mt_rand float args), imagerectangle, noise mt_rand.

Verified: 8x captcha generate with error_reporting(E_ALL) → 0 deprecation, output unchanged.

+++ securimage.php
@@ -60,6 +60,7 @@
  * @author     Drew Phillips <drew@drew-phillips.com>
  *
  */
+#[\AllowDynamicProperties]
 class Securimage
 {
     // All of the public variables below are securimage options
@@ -1935,7 +1936,7 @@ class Securimage
 
             if ($this->use_random_boxes && strlen(trim($char)) && mt_rand(1,100) % 5 == 0) {
                 imagesetthickness($im, 3);
-                imagerectangle($im, $x, $y - $dim[1] + $dim[2], $x + $dim[0], $y + $dim[2], $this->gdtextcolor);
+                imagerectangle($im, (int)$x, (int)($y - $dim[1] + $dim[2]), (int)($x + $dim[0]), (int)($y + $dim[2]), $this->gdtextcolor);
             }
 
             if ($c == ' ') {
@@ -1978,17 +1979,17 @@ class Securimage
         $amp      = array(); // amplitude
         $x        = ($this->image_width / 4); // lowest x coordinate of a pole
         $maxX     = $this->image_width - $x;  // maximum x coordinate of a pole
-        $dx       = mt_rand($x / 10, $x);     // horizontal distance between poles
+        $dx       = mt_rand((int)($x / 10), (int)$x);     // horizontal distance between poles
         $y        = mt_rand(20, $this->image_height - 20);  // random y coord
-        $dy       = mt_rand(20, $this->image_height * 0.7); // y distance
+        $dy       = mt_rand(20, (int)($this->image_height * 0.7)); // y distance
         $minY     = 20;                                     // minimum y coordinate
         $maxY     = $this->image_height - 20;               // maximum y cooddinate
 
         // make array of poles AKA attractor points
         for ($i = 0; $i < $numpoles; ++ $i) {
-            $px[$i]  = ($x + ($dx * $i)) % $maxX;
-            $py[$i]  = ($y + ($dy * $i)) % $maxY + $minY;
-            $rad[$i] = mt_rand($this->image_height * 0.4, $this->image_height * 0.8);
+            $px[$i]  = ((int)($x + ($dx * $i))) % (int)$maxX;
+            $py[$i]  = ((int)($y + ($dy * $i))) % $maxY + $minY;
+            $rad[$i] = mt_rand((int)($this->image_height * 0.4), (int)($this->image_height * 0.8));
             $tmp     = ((- $this->frand()) * 0.15) - .15;
             $amp[$i] = $this->perturbation * $tmp;
         }
@@ -2021,7 +2022,7 @@ class Securimage
                 $x *= $this->iscale;
                 $y *= $this->iscale;
                 if ($x >= 0 && $x < $width2 && $y >= 0 && $y < $height2) {
-                    $c = imagecolorat($this->tmpimg, $x, $y);
+                    $c = imagecolorat($this->tmpimg, (int)$x, (int)$y);
                 }
                 if ($c != $bgCol) { // only copy pixels of letters to preserve any background image
                     imagesetpixel($this->im, $ix, $iy, $c);
@@ -2038,11 +2039,11 @@ class Securimage
         for ($line = 0; $line < $this->num_lines; ++ $line) {
             $x = $this->image_width * (1 + $line) / ($this->num_lines + 1);
             $x += (0.5 - $this->frand()) * $this->image_width / $this->num_lines;
-            $y = mt_rand($this->image_height * 0.1, $this->image_height * 0.9);
+            $y = mt_rand((int)($this->image_height * 0.1), (int)($this->image_height * 0.9));
 
             $theta = ($this->frand() - 0.5) * M_PI * 0.33;
             $w = $this->image_width;
-            $len = mt_rand($w * 0.4, $w * 0.7);
+            $len = mt_rand((int)($w * 0.4), (int)($w * 0.7));
             $lwid = mt_rand(0, 2);
 
             $k = $this->frand() * 0.6 + 0.2;
@@ -2062,7 +2063,7 @@ class Securimage
             for ($i = 0; $i < $n; ++ $i) {
                 $x = $x0 + $i * $dx + $amp * $dy * sin($k * $i * $step + $phi);
                 $y = $y0 + $i * $dy - $amp * $dx * sin($k * $i * $step + $phi);
-                imagefilledrectangle($this->im, $x, $y, $x + $lwid, $y + $lwid, $this->gdlinecolor);
+                imagefilledrectangle($this->im, (int)$x, (int)$y, (int)($x + $lwid), (int)($y + $lwid), $this->gdlinecolor);
             }
         }
     }
